datetime: Move time zone to date time module
- It is more appropriate to have the time zone setting in date time
  module.
