+++ actions/timezone-change
@@ -25,6 +25,6 @@ if [ -e "$tzpath" ] ; then
     echo "$zonename" > /etc/timezone
     exit 0
 else
-    echo "timezone not valid" 1>&2
+    echo "Time zone not valid" 1>&2
     exit 1
 fi
